Define boot GPIO settings for RevPi Core 2022
Drive strength for bank 0 was previously set to 8mA by the Foundation's
original dt-blob.bin on the UART pins.  For bank 1, it was set to 8mA on
RevPi products to achieve better spi0 signal quality.  Hence both banks
use 8mA here.  It is assumed that 8mA is needed on bank 1 even on the
Core SE (which doesn't use spi0) to ensure decent signal quality for
i2c1.

All connected GPIOs are set up here, even though for some of them, the
default settings already reflect the desired state.  The default state
appears to be that "fsel" of all GPIOs is "input" and the default "pull"
is as documented on page 102f of the BCM2835 ARM Peripherals datasheet:

https://www.raspberrypi.org/app/uploads/2012/02/BCM2835-ARM-Peripherals.pdf

As a rule, no internal pull is configured here if an external pull is
present.

Note that the HAT EEPROM does not allow setting the initial polarity of
"output" GPIOs.  As a somewhat unsatisfactory workaround, some GPIOs are
set to "input" and their "pull" is set to the desired initial polarity.
Here's a list of the affected GPIOs:

 6  LED.ARED
16  LED.PWRRED
17  PB_RS485.TX_EN
30  LED.AGRN
32  LED.BGRN
33  LED.BRED
41  PB_RS485.TERM

For PB_RS485.TX_EN, it was considered to let the UART control the pin
via the Modem Control Register (fsel "alt3" and pull "none").  However
that would require a kernel patch for earlycon / earlyprintk code to
pick up RS-485 settings from the kernel command line or device tree
and to toggle the RTS bit upon sending.  Until we get to that, pull the
GPIO high (thus enabling RS-485 transmission).  The result is that the
Compute Module occupies the PiBridge RS-485 bus on boot, but that
shouldn't matter for the attached I/O modules and gateways.

On the Core SE, the following pins are not connected and thus not
configured vis-à-vis the Core 3+ and Core S:

12  PB_SPI.INT_RE
13  PB_SPI.INT_LI
20  PB_SPI.RST_RE
21  PB_SPI.RST_LI
35  PB_SPI.CS_RE
36  PB_SPI.CS_LI
37  PB_SPI.MISO
38  PB_SPI.MOSI_CM
39  PB_SPI.CLK_CM

Signed-off-by: Lukas Wunner <lukas@wunner.de>