fix: use SE specific config for SE products
Commit 9fd1e16 ("feat: add Core SE for screwable cooling") introduced
new product revisions for screwable heat sinks for both Core S and Core
SE's. These are different in the pin configuration for, where Core S'
have pins 12 and 13 listed, but Core SE's don't. This was overlooked.

Introduce a new template revpi-core-se-2022.json for Core SE products
where the pins 12 and 13 aren't listed in the gpiobanks.

Fixes: 9fd1e16 ("feat: add Core SE for screwable cooling")
Signed-off-by: Thomas Böhler <t.boehler@kunbus.com>